refactor(ui): extract SharePanel from ShareDialog
Extract the inner share controls + handlers into a new presentational
SharePanel component with no Dialog/Modal wrapper. ShareDialog becomes a
thin Dialog wrapper delegating to SharePanel, preserving its public props
so standalone call sites (Albums page) are unaffected.
